AI/requirements.txt
python-chess
fire
flask
flask-restx
numpy
torch
